HTTP request changed to produce JSON at all times.
This is temporary - will eventually need to return using the type requested by the caller,
but want to move to a data return rather than a HTML return (e.g. 404), to start that
progression.

Basic checks have been added to verify teh URL prior to starting more intensive processing.

The context now returns the parameters available, and the namespace the set of namespaces with data available based on the meta-type